Fixed nfa4 RE answer
I was just thinking about it this morning and realized that I did the wrong regular expression for nfa4. My logic was correct in that it should have given one or more "0"s. However, I attempted to display this by setting 0 to the power of +. This should indicate one or more zeros, but in retrospect this seems to be the incorrect format for this case, and a 00* seemed more appropriate.